Various minor sign/casting fixes

+++ fs2_open/code/mission/missionparse.cpp
@@ -9,13 +9,16 @@
 
 /*
  * $Logfile: /Freespace2/code/Mission/MissionParse.cpp $
- * $Revision: 2.178.2.42 $
- * $Date: 2007-11-21 07:27:47 $
- * $Author: Goober5000 $
+ * $Revision: 2.178.2.43 $
+ * $Date: 2007-11-24 04:33:49 $
+ * $Author: wmcoolmon $
  *
  * main upper level code for parsing stuff
  *
  * $Log: not supported by cvs2svn $
+ * Revision 2.178.2.42  2007/11/21 07:27:47  Goober5000
+ * add Wing Commander Saga's fiction viewer
+ *
  * Revision 2.178.2.41  2007/11/20 04:58:22  Goober5000
  * fix ship-type-destroyed
  *
@@ -4167,7 +4170,7 @@ p_object *mission_parse_get_parse_object(ushort net_signature)
 	int i;
 
 	// look for original ships
-	for (i = 0; i < Parse_objects.size(); i++)
+	for (i = 0; i < (int)Parse_objects.size(); i++)
 		if(Parse_objects[i].net_signature == net_signature)
 			return &Parse_objects[i];
 
@@ -4181,7 +4184,7 @@ p_object *mission_parse_get_parse_object(char *name)
 	int i;
 
 	// look for original ships
-	for (i = 0; i < Parse_objects.size(); i++)
+	for (i = 0; i < (int)Parse_objects.size(); i++)
 		if(!stricmp(Parse_objects[i].name, name))
 			return &Parse_objects[i];
 
@@ -4347,7 +4350,7 @@ int parse_wing_create_ships( wing *wingp, int num_to_create, int force, int spec
 
 	// Goober5000 - we have to do this via the array because we have no guarantee we'll be able to iterate along the list
 	// (since created objects plus anything they're docked to will be removed from it)
-	for (i = 0; i < Parse_objects.size(); i++)
+	for (i = 0; i < (int)Parse_objects.size(); i++)
 	{
 		int index;
 		ai_info *aip;
@@ -4828,7 +4831,7 @@ void parse_wing(mission *pm)
 		// everything is still only in the parse array at this point (in both FRED and FS2)
 
 		// find the parse object and assign it the wing number
-		for (j = 0; j < Parse_objects.size(); j++)
+		for (j = 0; j < (int)Parse_objects.size(); j++)
 		{
 			p_object *p_objp = &Parse_objects[j];
 
@@ -4928,7 +4931,7 @@ void post_process_path_stuff()
 	wing *wingp;
 
 	// take care of parse objects (ships)
-	for (i = 0; i < Parse_objects.size(); i++)
+	for (i = 0; i < (int)Parse_objects.size(); i++)
 	{
 		pobjp = &Parse_objects[i];
 
@@ -4963,7 +4966,7 @@ void post_process_ships_wings()
 	// Goober5000 - now create all objects that we can.  This must be done before any ship stuff
 	// but can't be done until the dock references are resolved.  This was originally done
 	// in parse_object().
-	for (i = 0; i < Parse_objects.size(); i++)
+	for (i = 0; i < (int)Parse_objects.size(); i++)
 	{
 		mission_parse_maybe_create_parse_object(&Parse_objects[i]);
 	}
@@ -4978,7 +4981,7 @@ void post_process_ships_wings()
 	if (Fred_running)
 	{
 		// even though the ships are already created, only the parse objects know the wing info
-		for (i = 0; i < Parse_objects.size(); i++)
+		for (i = 0; i < (int)Parse_objects.size(); i++)
 		{
 			p_object *p_objp = &Parse_objects[i];
 
@@ -6131,7 +6134,7 @@ void mission_parse_close()
 	}
 
 	// free parse object dock lists
-	for (int i = 0; i < Parse_objects.size(); i++)
+	for (int i = 0; i < (int)Parse_objects.size(); i++)
 	{
 		dock_free_instances(&Parse_objects[i]);
 	}
@@ -6307,7 +6310,7 @@ void parse_object_clear_handled_flag_helper(p_object *pobjp, p_dock_function_inf
 void parse_object_clear_all_handled_flags()
 {
 	// clear flag for all ships
-	for (int i = 0; i < Parse_objects.size(); i++)
+	for (int i = 0; i < (int)Parse_objects.size(); i++)
 	{
 		p_object *pobjp = &Parse_objects[i];
 		p_dock_function_info dfi;
@@ -6355,7 +6358,7 @@ void mission_parse_set_up_initial_docks()
 	}
 
 	// now resolve the leader of each tree
-	for (i = 0; i < Parse_objects.size(); i++)
+	for (i = 0; i < (int)Parse_objects.size(); i++)
 	{
 		p_object *pobjp = &Parse_objects[i];
 		p_dock_function_info dfi;
@@ -6882,7 +6885,7 @@ void mission_eval_arrivals()
 	// check the arrival list
 	// Goober5000 - we can't run through the list the usual way because we might
 	// remove a bunch of objects and completely screw up the list linkage
-	for (i = 0; i < Parse_objects.size(); i++)
+	for (i = 0; i < (int)Parse_objects.size(); i++)
 	{
 		p_object *pobjp = &Parse_objects[i];
